Override shutdown method on _StandaloneSSLConnection object to strip "how" argument.
Normal socket has "how" argument but OpenSSL.SSL.Connection doesn't. To avoid
an error, insert a function to remove the argument.

Bonus:
Quit if --tls-client-auth is specified but ssl module is not available.
We log that at critical level now but forgot to exit.
Review URL: https://codereview.appspot.com/7374048